Extract PurchaseMeta into new component
