+++ client/me/purchases/utils.js
@@ -8,7 +8,14 @@ import { get } from 'lodash';
  * Internal dependencies
  */
 import analytics from 'lib/analytics';
+import config from 'config';
 import paths from './paths';
+import {
+	isExpired,
+	isIncludedWithPlan,
+	isOneTimePurchase,
+	isPaidWithCreditCard
+} from 'lib/purchases';
 
 // TODO: Remove these property-masking functions in favor of accessing the props directly
 function getPurchase( props ) {
@@ -84,6 +91,22 @@ function enrichedSurveyData( surveyData, moment, site, purchase ) {
 	);
 }
 
+function canEditPaymentDetails( purchase ) {
+	if ( ! config.isEnabled( 'upgrades/credit-cards' ) ) {
+		return false;
+	}
+	return ! isExpired( purchase ) && ! isOneTimePurchase( purchase ) && ! isIncludedWithPlan( purchase );
+}
+
+function getEditCardDetailsPath( site, purchase ) {
+	if ( isPaidWithCreditCard( purchase ) ) {
+		const { payment: { creditCard } } = purchase;
+
+		return paths.editCardDetails( site.slug, purchase.id, creditCard.id );
+	}
+	return paths.addCardDetails( site.slug, purchase.id );
+}
+
 export {
 	getPurchase,
 	getSelectedSite,
@@ -93,4 +116,6 @@ export {
 	isDataLoading,
 	recordPageView,
 	enrichedSurveyData,
+	canEditPaymentDetails,
+	getEditCardDetailsPath,
 };
